requirements.txt
Flask==2.3.2
PyJWT==2.8.0
flask-cors==4.0.0
